refactored civitdl/batch directory
