+++ src/civitdl/batch/batch_download.py
@@ -2,7 +2,7 @@ import time
 import traceback
 from typing import List
 
-from ._get_model import download_model
+from ._model import Model
 
 from helpers.core.utils import Styler, APIException, get_version, print_exc, print_verbose, run_verbose, sprint
 from helpers.sourcemanager import SourceManager
@@ -27,11 +27,8 @@ def batch_download(source_strings: List[str], rootdir: str, batchOptions: BatchO
         iter = 0
         while True:
             try:
-                download_model(
-                    id=id,
-                    dst_root_path=rootdir,
-                    batchOptions=batchOptions
-                )
+                model = Model(id=id, dst_root_path=rootdir,
+                              batchOptions=batchOptions).download()
                 _pause(batchOptions.pause_time)
                 break
             except Exception as e:
@@ -40,8 +37,8 @@ def batch_download(source_strings: List[str], rootdir: str, batchOptions: BatchO
                 print_exc(e, '\n')
                 sprint('---------')
                 _pause(batchOptions.pause_time)
-                if not isinstance(e, APIException):
-                    break
+                # if not isinstance(e, APIException):
+                #     break
                 if iter < batchOptions.retry_count:
                     sprint(Styler.stylize(
                         'Retrying to download the current model...', color='info'))
